Laravel 8 Support: Batching Requires UUID
Since Laravel 8 landed with some modifications to the
DatabaseUuidFailedJobProvider

[link](https://github.com/illuminate/queue/blame/d70c2df8ae9cd52db24c4f09333c36221b092b6d/Failed/DatabaseUuidFailedJobProvider.php#L58)

I was receiving some exceptions that would prevent jobs from being
written to the database (exception to be attached in PR comment)

This basically shuffles the MessageId, which I can confirm is not
rejected by Laravel, into the UUID in Laravel's terms.

Fair warning I'm not clear on the implications of this change, but it
does at least succeed in writing failed jobs to the database rather than
crashing.